[feature](score) Add score function (#52113)

+++ be/src/olap/rowset/segment_v2/inverted_index_iterator.cpp
@@ -48,7 +48,7 @@ Status InvertedIndexIterator::read_from_index(const IndexParam& param) {
             runtime_state->query_options().inverted_index_skip_threshold < 100) {
             auto query_bkd_limit_percent =
                     runtime_state->query_options().inverted_index_skip_threshold;
-            uint32_t hit_count = 0;
+            size_t hit_count = 0;
             RETURN_IF_ERROR(try_read_from_inverted_index(i_param->column_name, i_param->query_value,
                                                          i_param->query_type, &hit_count));
             if (hit_count > i_param->num_rows * query_bkd_limit_percent / 100) {
